lsp: fix warning on macOS
gint64 is "long long" on macOS and "long" on linux which leads to a
warning in the format specifier.

+++ lsp/src/lsp-log.c
@@ -110,7 +110,7 @@ void lsp_log(LspLogInfo log, LspLogType type, const gchar *method, GVariant *par
 	if (req_time)
 	{
 		GTimeSpan delta = g_date_time_difference(time, req_time);
-		delta_str = g_strdup_printf(" (%ld ms)", delta / 1000);
+		delta_str = g_strdup_printf(" (%ld ms)", (long)(delta / 1000));
 	}
 	else
 		delta_str = g_strdup("");
